+++ src/api/authFactor.ts
@@ -1,31 +1,45 @@
-import { type AuthFactor, urls } from "codeforlife/api"
+import { type AuthFactor, type User, urls } from "codeforlife/api"
 import {
   type CreateArg,
   type CreateResult,
   type DestroyArg,
   type DestroyResult,
+  type ListArg,
+  type ListResult,
   buildUrl,
   tagData,
 } from "codeforlife/utils/api"
 import getReadAuthFactorEndpoints, {
   AUTH_FACTOR_TAG,
-  type ListAuthFactorsArg,
-  type ListAuthFactorsResult,
 } from "codeforlife/api/endpoints/authFactor"
 
 import api from "."
 
-export type { ListAuthFactorsArg, ListAuthFactorsResult }
+export type ListAuthFactorsResult = ListResult<AuthFactor, "type">
+export type ListAuthFactorsArg = ListArg<{
+  user: User["id"]
+  type: AuthFactor["type"]
+}>
 
 export type CreateAuthFactorResult = CreateResult<AuthFactor>
-export type CreateAuthFactorArg = CreateArg<AuthFactor, "type">
+export type CreateAuthFactorArg = CreateArg<AuthFactor, "type"> & {
+  otp?: string
+}
 
 export type DestroyAuthFactorResult = DestroyResult
 export type DestroyAuthFactorArg = DestroyArg<AuthFactor>
 
+export type GetOtpSecretResult = {
+  secret: string
+  provisioning_uri: string
+}
+export type GetOtpSecretArg = null
+
 const authFactorApi = api.injectEndpoints({
   endpoints: build => ({
-    ...getReadAuthFactorEndpoints(build),
+    ...getReadAuthFactorEndpoints<ListAuthFactorsResult, ListAuthFactorsArg>(
+      build,
+    ),
     createAuthFactor: build.mutation<
       CreateAuthFactorResult,
       CreateAuthFactorArg
@@ -35,7 +49,9 @@ const authFactorApi = api.injectEndpoints({
         method: "POST",
         body,
       }),
-      invalidatesTags: tagData(AUTH_FACTOR_TAG, { includeListTag: true }),
+      // NOTE: Intentionally not invalidating tags to show success sub-view.
+      // TODO: assess whether sub-view can be replaced with notification.
+      // invalidatesTags: tagData(AUTH_FACTOR_TAG, { includeListTag: true }),
     }),
     destroyAuthFactor: build.mutation<
       DestroyAuthFactorResult,
@@ -47,6 +63,12 @@ const authFactorApi = api.injectEndpoints({
       }),
       invalidatesTags: tagData(AUTH_FACTOR_TAG, { includeListTag: true }),
     }),
+    getOtpSecret: build.query<GetOtpSecretResult, GetOtpSecretArg>({
+      query: () => ({
+        url: urls.authFactor.list + "get-otp-secret/",
+        method: "GET",
+      }),
+    }),
   }),
 })
 
@@ -56,4 +78,5 @@ export const {
   useDestroyAuthFactorMutation,
   useListAuthFactorsQuery,
   useLazyListAuthFactorsQuery,
+  useGetOtpSecretQuery,
 } = authFactorApi